apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.136.1",
    "psycopg>=3.3.3",
    "pydantic-settings>=2.14.0",
    "sqlalchemy>=2.0.49",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
